Department store management
This program is basically a billing system that we often see in our supermarkets, groceries etc., Here I have simplified the work of the billing workers by the effective use of C program. My first billing app consist of main menu which comprises of various functions like Goods entry, their edits, to search for a particular entry and atlast to calculate the bill. The main moto is to make the billing progress simple and easy to the workers and also to the customers who wait in long Q's.